fix(omp): soft-stop nested task teardown without killing parent identity
OMP task subagents share the parent process and were binding/stopping the
parent on session_shutdown. Soft-finalize on owner shutdown, skip nested
bind via a process-wide identity latch, recover bindings from
HCOM_INSTANCE_NAME, and split the omp hook module under 1k lines.
