+++ src/hooks/omp/handlers.rs
@@ -1,0 +1,365 @@
+//! Omp Coding Agent hook handlers — argv-based lifecycle plus TypeScript plugin.
+
+use std::time::Instant;
+
+use serde_json::Value;
+
+use crate::bootstrap;
+use crate::db::HcomDb;
+use crate::instance_binding;
+use crate::instance_lifecycle as lifecycle;
+use crate::instances;
+use crate::log::{log_error, log_info};
+use crate::shared::ST_LISTENING;
+use crate::shared::context::HcomContext;
+
+use crate::hooks::common;
+use crate::hooks::common::finalize_session;
+
+fn parse_flag(argv: &[String], flag: &str) -> Option<String> {
+    argv.iter()
+        .position(|a| a == flag)
+        .and_then(|i| argv.get(i + 1))
+        .cloned()
+}
+
+fn has_flag(argv: &[String], flag: &str) -> bool {
+    argv.iter().any(|a| a == flag)
+}
+
+pub(crate) fn upsert_plugin_notify_endpoint(db: &HcomDb, instance_name: &str, port: u16) {
+    if let Err(e) = db.upsert_notify_endpoint(instance_name, "plugin", port) {
+        log_error(
+            "native",
+            "omp.register_notify_fail",
+            &format!(
+                "Failed to register plugin notify port for {}: {}",
+                instance_name, e
+            ),
+        );
+        return;
+    }
+
+    crate::notify::wake(db, instance_name, crate::notify::WakeKind::DELIVERY_LOOPS);
+}
+
+fn initialize_last_event_id(db: &HcomDb, instance_name: &str) {
+    if let Ok(Some(existing)) = db.get_instance_full(instance_name)
+        && existing.last_event_id == 0
+    {
+        let launch_event_id: Option<i64> = std::env::var("HCOM_LAUNCH_EVENT_ID")
+            .ok()
+            .and_then(|s| s.parse().ok());
+        let current_max = db.get_last_event_id();
+        let new_id = match launch_event_id {
+            Some(lei) if lei <= current_max => lei,
+            _ => current_max,
+        };
+        let mut updates = serde_json::Map::new();
+        updates.insert("last_event_id".into(), serde_json::json!(new_id));
+        instances::update_instance_position(db, instance_name, &updates);
+    }
+}
+
+fn instance_name_from_env(ctx: &HcomContext) -> Option<String> {
+    ctx.raw_env
+        .get("HCOM_INSTANCE_NAME")
+        .or_else(|| ctx.raw_env.get("HCOM_NAME"))
+        .filter(|s| !s.is_empty())
+        .cloned()
+}
+
+fn bootstrap_for(ctx: &HcomContext, db: &HcomDb, instance_name: &str) -> String {
+    let tag = db
+        .get_instance_full(instance_name)
+        .ok()
+        .flatten()
+        .and_then(|d| d.tag.clone())
+        .unwrap_or_default();
+    let hcom_config = crate::config::HcomConfig::load(None).unwrap_or_default();
+    let relay_enabled = crate::relay::is_relay_enabled(&hcom_config);
+    let effective_tag = if tag.is_empty() {
+        &hcom_config.tag
+    } else {
+        &tag
+    };
+    bootstrap::get_bootstrap(
+        db,
+        &ctx.hcom_dir,
+        instance_name,
+        "omp",
+        ctx.is_background,
+        ctx.is_launched,
+        &ctx.notes,
+        effective_tag,
+        relay_enabled,
+        ctx.background_name.as_deref(),
+    )
+}
+
+pub(crate) fn handle_start(ctx: &HcomContext, db: &HcomDb, argv: &[String]) -> (i32, String) {
+    // Plugin RPC returns JSON errors on exit 0 so the extension can handle
+    // setup failures without Pi treating the hook itself as failed.
+    let session_id = match parse_flag(argv, "--session-id") {
+        Some(sid) => sid,
+        None => return (0, r#"{"error":"Missing --session-id"}"#.to_string()),
+    };
+    let transcript_path = parse_flag(argv, "--transcript-path");
+    let cwd = parse_flag(argv, "--cwd");
+    let notify_port: Option<u16> = parse_flag(argv, "--notify-port").and_then(|s| s.parse().ok());
+
+    let process_id = match &ctx.process_id {
+        Some(pid) => pid.clone(),
+        None => return (0, r#"{"error":"HCOM_PROCESS_ID not set"}"#.to_string()),
+    };
+
+    let instance_name =
+        match instance_binding::bind_session_to_process(db, &session_id, Some(&process_id)) {
+            Some(name) => name,
+            None => match instance_name_from_env(ctx).and_then(|name| {
+                instance_binding::recover_process_binding_for_instance(
+                    db,
+                    &name,
+                    &session_id,
+                    &process_id,
+                )
+            }) {
+                Some(name) => name,
+                None => {
+                    return (
+                        0,
+                        r#"{"error":"No instance bound to this process"}"#.to_string(),
+                    );
+                }
+            },
+        };
+
+    initialize_last_event_id(db, &instance_name);
+    lifecycle::set_status(
+        db,
+        &instance_name,
+        ST_LISTENING,
+        "start",
+        Default::default(),
+    );
+    instance_binding::capture_and_store_launch_context(db, &instance_name);
+
+    let mut updates = serde_json::Map::new();
+    updates.insert("tool".into(), serde_json::json!("omp"));
+    updates.insert("session_id".into(), serde_json::json!(&session_id));
+    if let Some(path) = transcript_path.as_ref().filter(|p| !p.is_empty()) {
+        updates.insert("transcript_path".into(), serde_json::json!(path));
+    }
+    let cwd_value = cwd
+        .as_deref()
+        .filter(|p| !p.is_empty())
+        .or_else(|| ctx.cwd.to_str());
+    if let Some(cwd) = cwd_value {
+        updates.insert("directory".into(), serde_json::json!(cwd));
+    }
+    instances::update_instance_position(db, &instance_name, &updates);
+    if let Some(port) = notify_port {
+        upsert_plugin_notify_endpoint(db, &instance_name, port);
+    }
+    log_info(
+        "hooks",
+        "omp-start.bind",
+        &format!("instance={} session_id={}", instance_name, session_id),
+    );
+    crate::relay::worker::ensure_worker(true);
+
+    let response = serde_json::json!({
+        "name": instance_name,
+        "session_id": session_id,
+        "bootstrap": bootstrap_for(ctx, db, &instance_name),
+    });
+    (0, response.to_string())
+}
+
+pub(crate) fn handle_status(db: &HcomDb, argv: &[String]) -> (i32, String) {
+    let name = match parse_flag(argv, "--name") {
+        Some(n) => n,
+        None => return (0, r#"{"error":"Missing --name or --status"}"#.to_string()),
+    };
+    let status = match parse_flag(argv, "--status") {
+        Some(s) => s,
+        None => return (0, r#"{"error":"Missing --name or --status"}"#.to_string()),
+    };
+    let context = parse_flag(argv, "--context").unwrap_or_default();
+    let detail = parse_flag(argv, "--detail").unwrap_or_default();
+    let was_listening = db
+        .get_instance_full(&name)
+        .ok()
+        .flatten()
+        .is_some_and(|inst| inst.status == ST_LISTENING);
+
+    lifecycle::set_status(
+        db,
+        &name,
+        &status,
+        &context,
+        lifecycle::StatusUpdate {
+            detail: &detail,
+            ..Default::default()
+        },
+    );
+    if status == ST_LISTENING && !was_listening {
+        crate::notify::wake(db, &name, &[]);
+    }
+    (0, r#"{"ok":true}"#.to_string())
+}
+
+fn handle_read(db: &HcomDb, argv: &[String]) -> (i32, String) {
+    let name = match parse_flag(argv, "--name") {
+        Some(n) => n,
+        None => return (0, r#"{"error":"Missing --name"}"#.to_string()),
+    };
+    let format_mode = has_flag(argv, "--format");
+    let check_mode = has_flag(argv, "--check");
+    let ack_mode = has_flag(argv, "--ack");
+
+    let raw_messages = db.get_unread_messages(&name);
+    let messages: Vec<Value> = raw_messages.iter().map(common::message_to_value).collect();
+
+    if format_mode {
+        if messages.is_empty() {
+            return (0, String::new());
+        }
+        let deliver = common::limit_delivery_messages(&messages);
+        return (
+            0,
+            common::format_messages_json_for_instance(db, &deliver, &name),
+        );
+    }
+    if ack_mode {
+        if let Some(up_to) = parse_flag(argv, "--up-to") {
+            let Ok(ack_id) = up_to.parse::<i64>() else {
+                return (
+                    0,
+                    serde_json::json!({"error": format!("Invalid --up-to: {}", up_to)}).to_string(),
+                );
+            };
+            let mut updates = serde_json::Map::new();
+            updates.insert("last_event_id".into(), serde_json::json!(ack_id));
+            instances::update_instance_position(db, &name, &updates);
+            return (0, serde_json::json!({"acked_to": ack_id}).to_string());
+        }
+        if messages.is_empty() {
+            return (0, r#"{"acked":0}"#.to_string());
+        }
+        let ack_id = messages
+            .iter()
+            .filter_map(|m| m.get("event_id").and_then(|v| v.as_i64()))
+            .max()
+            .filter(|id| *id > 0)
+            .unwrap_or_else(|| db.get_last_event_id());
+        if ack_id > 0 {
+            let mut updates = serde_json::Map::new();
+            updates.insert("last_event_id".into(), serde_json::json!(ack_id));
+            instances::update_instance_position(db, &name, &updates);
+        }
+        return (0, serde_json::json!({"acked": messages.len()}).to_string());
+    }
+    if check_mode {
+        return (
+            0,
+            if messages.is_empty() { "false" } else { "true" }.to_string(),
+        );
+    }
+    (
+        0,
+        serde_json::to_string(&messages).unwrap_or_else(|_| "[]".to_string()),
+    )
+}
+
+fn handle_beforetool(db: &HcomDb, argv: &[String]) -> (i32, String) {
+    let name = match parse_flag(argv, "--name") {
+        Some(n) => n,
+        None => return (0, r#"{"decision":"allow"}"#.to_string()),
+    };
+    let tool_name = parse_flag(argv, "--tool").unwrap_or_default();
+    let input = parse_flag(argv, "--input-json")
+        .and_then(|raw| serde_json::from_str::<Value>(&raw).ok())
+        .unwrap_or_else(|| serde_json::json!({}));
+    if !tool_name.is_empty() {
+        common::update_tool_status(db, &name, "omp", &tool_name, &input);
+    }
+    (0, r#"{"decision":"allow"}"#.to_string())
+}
+
+pub(crate) fn handle_stop(db: &HcomDb, argv: &[String]) -> (i32, String) {
+    let name = match parse_flag(argv, "--name") {
+        Some(n) => n,
+        None => return (0, r#"{"error":"Missing --name"}"#.to_string()),
+    };
+    let reason = parse_flag(argv, "--reason").unwrap_or_else(|| "unknown".to_string());
+    if has_flag(argv, "--soft") {
+        common::soft_finalize_session(db, &name, &reason, None);
+        (0, r#"{"ok":true,"soft":true}"#.to_string())
+    } else {
+        finalize_session(db, &name, &reason, None);
+        (0, r#"{"ok":true}"#.to_string())
+    }
+}
+
+pub fn dispatch_omp_hook(hook_name: &str, argv: &[String]) -> (i32, String) {
+    let start = Instant::now();
+    let ctx = HcomContext::from_os();
+    crate::paths::ensure_hcom_directories_at(&ctx.hcom_dir);
+    let db = match HcomDb::open() {
+        Ok(db) => db,
+        Err(e) => {
+            log_error(
+                "hooks",
+                "hook.error",
+                &format!("hook={} op=db_open err={}", hook_name, e),
+            );
+            return (
+                0,
+                serde_json::json!({"error": format!("DB open failed: {}", e)}).to_string(),
+            );
+        }
+    };
+    if !common::hook_gate_check(&ctx, &db) {
+        return (0, String::new());
+    }
+    let handler_argv: Vec<String> = if !argv.is_empty() && argv[0] == hook_name {
+        argv[1..].to_vec()
+    } else {
+        argv.to_vec()
+    };
+    let hook_name_owned = hook_name.to_string();
+    let handler_start = Instant::now();
+    let (exit_code, output) = common::dispatch_with_panic_guard(
+        "omp",
+        &hook_name_owned,
+        (
+            0,
+            serde_json::json!({"error": "internal panic"}).to_string(),
+        ),
+        || match hook_name_owned.as_str() {
+            "omp-start" => handle_start(&ctx, &db, &handler_argv),
+            "omp-status" => handle_status(&db, &handler_argv),
+            "omp-read" => handle_read(&db, &handler_argv),
+            "omp-beforetool" => handle_beforetool(&db, &handler_argv),
+            "omp-stop" => handle_stop(&db, &handler_argv),
+            _ => (
+                0,
+                serde_json::json!({"error": format!("Unknown Omp hook: {}", hook_name_owned)})
+                    .to_string(),
+            ),
+        },
+    );
+    log_info(
+        "hooks",
+        "omp.dispatch.timing",
+        &format!(
+            "hook={} handler_ms={:.2} total_ms={:.2} exit_code={}",
+            hook_name,
+            handler_start.elapsed().as_secs_f64() * 1000.0,
+            start.elapsed().as_secs_f64() * 1000.0,
+            exit_code
+        ),
+    );
+    (exit_code, output)
+}
